--- app/Contracts/Job/UpdateJobContract.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace App\Contracts\Job;
-
-use App\Models\JobInput;
-
-interface UpdateJobContract
-{
-    /**
-     * Update a JobInput.
-     *
-     * @param array $input
-     * @param JobInput $promotion
-     * @return bool
-     */
-    public function update(array $input, JobInput $promotion): bool;
-}
