--- app/Contracts/Job/UpdateJobResponseContract.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Contracts\Job;
-
-use Illuminate\Contracts\Support\Responsable;
-
-interface UpdateJobResponseContract extends Responsable
-{
-    //
-}
