--- app/Http/Responses/Job/UpdateJobResponse.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace App\Http\Responses\Job;
-
-use App\Contracts\Job\UpdateJobResponseContract;
-use Illuminate\Http\Request;
-use Symfony\Component\HttpFoundation\Response;
-
-class UpdateJobResponse implements UpdateJobResponseContract
-{
-    /**
-     * Create an HTTP response that represents the object.
-     *
-     * @param Request $request
-     * @return Response
-     */
-    public function toResponse($request): Response
-    {
-        return redirect()->to('jobs');
-    }
-}
