+++ resources/js/Layouts/AppLayout.vue
@@ -56,6 +56,9 @@ const stopImpersonation = () => {
 
                             <!-- Navigation Links -->
                             <div class="hidden space-x-8 sm:-my-px sm:ml-10 sm:flex">
+                                <JetNavLink class="dark:text-gray-200" :href="route('import.index')" :active="route().current('import.index')">
+                                    Import Data
+                                </JetNavLink>
                                 <JetNavLink class="dark:text-gray-200" :href="route('jobs.create')" :active="route().current('jobs.create')">
                                     Create Jobs
                                 </JetNavLink>
@@ -241,6 +244,9 @@ const stopImpersonation = () => {
                 <!-- Responsive Navigation Menu -->
                 <div :class="{'block': showingNavigationDropdown, 'hidden': ! showingNavigationDropdown}" class="sm:hidden">
                     <div class="pt-2 pb-3 space-y-1">
+                        <JetResponsiveNavLink class="dark:text-gray-200" :href="route('import.index')" :active="route().current('import.index')">
+                            Import Data
+                        </JetResponsiveNavLink>
                         <JetResponsiveNavLink class="dark:text-gray-200" :href="route('jobs.create')" :active="route().current('jobs.create')">
                             Create Jobs
                         </JetResponsiveNavLink>
